refactor(queries): modularize button, buttons, link, & links from text
