+++ src/queries/index.ts
@@ -1,7 +1,11 @@
 export * from './alt';
+export * from './button';
+export * from './buttons';
 export * from './element';
 export * from './focused';
 export * from './label';
+export * from './link';
+export * from './links';
 export * from './placeholder';
 export * from './testid';
 export * from './text';
